+++ Cask/remindian.rb
@@ -1,6 +1,6 @@
 cask "remindian" do
-  version "5.3.0"
-  sha256 "5c40f90518ec1be673f4078260ae19c0e6823b129c9963c6cb065be2f554025b"
+  version "5.4.0"
+  sha256 "dfbd5e2a0490cc3f4b3db7d9392437574f77d094a6603c4f67c8b183ed9f2130"
 
   url "https://github.com/Santofer/Remindian/releases/download/v#{version}/Remindian-v#{version}.dmg"
   name "Remindian"